rename: 변경된 cn 경로 적용
