+++ apps/web/src/components/button.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/lib/utils";
 
 export default function Button(props: React.ButtonHTMLAttributes<HTMLButtonElement>) {
   const { children, className, ...restProps } = props;
