+++ apps/web/src/components/query-tab-nav.tsx
@@ -4,7 +4,7 @@ import { Suspense } from "react";
 import Link from "next/link";
 import { useSearchParams } from "next/navigation";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/lib/utils";
 
 interface QueryTabNavProps {
   /**
